# docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateDetectionRequest:
    properties:
      confidenceThreshold:
        type: number
      device:
        type: string
      inferenceTimeMs:
        type: integer
      iouThreshold:
        type: number
      items:
        description: 检测项
        items:
          $ref: '#/definitions/models.DetectionItem'
        type: array
      modelName:
        type: string
      modelVersion:
        type: string
      processedFilename:
        type: string
      processedPath:
        type: string
      runId:
        description: 可选：运行ID，用于幂等
        type: string
      sourceFilename:
        type: string
      sourcePath:
        type: string
      summary:
        allOf:
        - $ref: '#/definitions/models.DetectionSummary'
        description: 汇总信息（是否有问题等）
    type: object
  handlers.DetectEntryRequest:
    properties:
      conf:
        type: number
      device:
        type: string
      imageId:
        type: string
      iou:
        type: number
      modelName:
        type: string
      modelVersion:
        type: string
      weights:
        type: string
    type: object
  handlers.StartDetectRequest:
    properties:
      conf:
        description: 置信度阈值，默认 0.25
        type: number
      device:
        description: 设备：cpu/cuda:0/mps，默认空（由YOLO自动选择）
        type: string
      iou:
        description: IoU 阈值，默认 0.45
        type: number
      modelName:
        description: 模型名称，默认 "best"
        type: string
      modelVersion:
        description: 模型版本，可选
        type: string
      weights:
        description: 模型权重文件路径或名称，默认由 room 配置覆盖
        type: string
    type: object
...
  models.BoundingBox:
    properties:
      height:
        type: number
      width:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
  models.DetectionItem:
    properties:
      bbox:
        $ref: '#/definitions/models.BoundingBox'
      class:
        type: string
      classId:
        type: integer
      confidence:
        type: number
      note:
        type: string
      polygon:
        items:
          $ref: '#/definitions/models.PolygonPoint'
        type: array
    type: object
  models.DetectionSummary:
    properties:
      avgScore:
        type: number
      hasIssue:
        type: boolean
      issueType:
        type: string
      objectCount:
        type: integer
    type: object
  models.PolygonPoint:
    properties:
      x:
        type: number
      "y":
        type: number
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@foreignscan.com
    name: API Support
    url: http://www.foreignscan.com/support
  description: ForeignScan后端API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForeignScan API
  version: "1.0"
paths:
  /detect:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.DetectEntryRequest'
      produces:
      - application/json
      responses:
        "202":
          description: 任务已启动，返回jobId与初始状态
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 兼容入口：前端直接传 imageId 触发单图推理
      tags:
      - detections
  /detect/jobs/{id}:
    delete:
      consumes:
      - application/json
      description: 取消任务（支持pending/running/parsing阶段），任务将标记为canceled
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 取消成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 未找到任务
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 任务不可取消（已结束）
          schema:
            additionalProperties: true
            type: object
      summary: 取消正在运行的检测任务
      tags:
      - detections
    get:
      consumes:
      - application/json
      description: 前端轮询或SSE，可获取任务进度/状态
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务状态
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 未找到任务
          schema:
            additionalProperties: true
            type: object
      summary: 查询推理任务状态
      tags:
      - detections
  /detect/jobs/{id}/stream:
    get:
      description: 通过SSE实时获取任务状态更新，直到任务结束
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE事件流
          schema:
            type: string
        "404":
          description: 未找到任务
          schema:
            additionalProperties: true
            type: object
      summary: 订阅检测任务实时进度（SSE）
      tags:
      - detections
  /detections:
    get:
      consumes:
      - application/json
      description: 支持按房间、点位、时间范围、是否有问题、类别等条件筛选检测运行
      parameters:
      - description: 房间ID
        in: query
        name: roomId
        type: string
      - description: 点位ID
        in: query
        name: pointId
        type: string
      - description: 是否存在问题
        in: query
        name: hasIssue
        type: boolean
      - description: 目标类别名称筛选
        in: query
        name: class
        type: string
      - description: 开始日期（YYYY-MM-DD）
        in: query
        name: start
        type: string
      - description: 结束日期（YYYY-MM-DD）
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功查询检测结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 查询检测结果
      tags:
      - detections
...
  /images/{id}/detect:
    post:
      consumes:
      - application/json
      description: 后端启动异步任务，调用YOLO并写入数据库（单张图片）
      parameters:
      - description: 图片ID
        in: path
        name: id
        required: true
        type: string
      - description: 推理配置，可选
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.StartDetectRequest'
      produces:
      - application/json
      responses:
        "202":
          description: 任务已启动，返回jobId与初始状态
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 前端触发单图异步推理
      tags:
      - detections
  /images/{id}/detections:
    get:
      consumes:
      - application/json
      description: 根据图片ID获取所有检测运行结果（按创建时间倒序）
      parameters:
      - description: 图片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取检测结果
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 图片不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取图片的检测结果列表
      tags:
      - detections
    post:
      consumes:
      - application/json
      description: YOLO推理完成后，向后端写入一次检测运行结果；支持RunID幂等插入，自动更新图片摘要字段
      parameters:
      - description: 图片ID
        in: path
        name: id
        required: true
        type: string
      - description: 检测结果请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDetectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 成功写入检测结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 图片不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 写入图片的检测结果
      tags:
      - detections
...
schemes:
- http
swagger: "2.0"